automatically update yarn.lock based on changes to package.json
This is basically just `yarn && git add yarn.lock && git commit`. We
hold @material-ui/core at 1.0.0 because this is what upstream does,
and moving it to 1.5.3 or whatever seriously breaks things. (It
expects React 16.) Holding @material-ui/core at 1.0.0 does not cause
vulnerability problems.

This yarn.lock has plenty of security vulnerabilities that will be
resolved in the following commits.