Version v7.7.7 RC (#8161)
* Version v7.7.7

* Update `@metamask/eth-ledger-bridge-keyring` (#8162)

* Update `@metamask/eth-ledger-bridge-keyring`

The Ledger keyring has been updated to ensure that any stale BIP44
accounts created prior to v7.7.6 of the extension are discarded when
the extension starts. Any attempts to sign with these accounts would
have failed; they needed to be re-added regardless.

* Update changelog

* Fix Ledger account index check for account zero (#8163)

Update Ledger keyring to fix bug when trying to sign with account 0

Co-authored-by: Mark Stacey <markjstacey@gmail.com>